Clean a little the code with bad instructions
